🚚 Move usecases
