+++ lib/ui/views/add_website/bloc/provider.dart
@@ -1,7 +1,7 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
 import 'dart:developer';
 
-import 'package:aeweb/domain/usecases/website/add_website.usecase.dart';
+import 'package:aeweb/domain/usecases/add_website.usecase.dart';
 import 'package:aeweb/ui/views/add_website/bloc/state.dart';
 import 'package:aeweb/util/certificate_util.dart';
 import 'package:aeweb/util/file_util.dart';
