+++ lib/ui/views/add_website/layouts/components/add_website_in_progress_popup.dart
@@ -1,6 +1,6 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
 import 'package:aeweb/application/websites.dart';
-import 'package:aeweb/domain/usecases/website/add_website.usecase.dart';
+import 'package:aeweb/domain/usecases/add_website.usecase.dart';
 import 'package:aeweb/ui/views/add_website/bloc/provider.dart';
 import 'package:aeweb/ui/views/add_website/layouts/components/add_website_circular_step_progress_indicator.dart';
 import 'package:aeweb/ui/views/util/components/countdown.dart';
@@ -139,17 +139,18 @@ class AddWebsiteInProgressPopup {
                                       top: 200,
                                     ),
                                     child: Card(
-                                        color: Colors.transparent,
-                                        clipBehavior: Clip.antiAlias,
-                                        elevation: 0,
-                                        margin: EdgeInsets.zero,
-                                        shape: RoundedRectangleBorder(
-                                          borderRadius: BorderRadius.only(
-                                            bottomLeft: Radius.circular(16),
-                                            bottomRight: Radius.circular(16),
-                                          ),
+                                      color: Colors.transparent,
+                                      clipBehavior: Clip.antiAlias,
+                                      elevation: 0,
+                                      margin: EdgeInsets.zero,
+                                      shape: RoundedRectangleBorder(
+                                        borderRadius: BorderRadius.only(
+                                          bottomLeft: Radius.circular(16),
+                                          bottomRight: Radius.circular(16),
                                         ),
-                                        child: aedappfm.PopupWaves()),
+                                      ),
+                                      child: aedappfm.PopupWaves(),
+                                    ),
                                   ),
                                   Padding(
                                     padding: const EdgeInsets.all(20),
