+++ lib/ui/views/update_certificate/bloc/provider.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:aeweb/domain/usecases/website/update_certificate.usecase.dart';
+import 'package:aeweb/domain/usecases/update_certificate.usecase.dart';
 import 'package:aeweb/ui/views/update_certificate/bloc/state.dart';
 import 'package:aeweb/util/certificate_util.dart';
 import 'package:archethic_dapp_framework_flutter/archethic_dapp_framework_flutter.dart'
