+++ lib/ui/views/update_website_sync/layouts/components/update_website_sync_comparison_list.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:aeweb/domain/usecases/website/sync_website.dart';
+import 'package:aeweb/domain/usecases/sync_website.dart';
 import 'package:aeweb/ui/views/update_website_sync/bloc/provider.dart';
 import 'package:archethic_dapp_framework_flutter/archethic_dapp_framework_flutter.dart'
     as aedappfm;
