+++ lib/ui/views/util/choose_path_sync_popup.dart
@@ -1,7 +1,7 @@
 import 'dart:developer';
 
-import 'package:aeweb/domain/usecases/website/read_website_version.usecase.dart';
-import 'package:aeweb/domain/usecases/website/sync_website.dart';
+import 'package:aeweb/domain/usecases/read_website_version.usecase.dart';
+import 'package:aeweb/domain/usecases/sync_website.dart';
 import 'package:aeweb/ui/views/update_website_sync/layouts/update_website_sync_sheet.dart';
 import 'package:aeweb/ui/views/util/warning_size_label.dart';
 import 'package:aeweb/util/file_util.dart';
